gemu: SS length is up to 256 (LL=0xff) - widen len past uint8_t
EXEC_SS computed len = (rL1 & 0xff) + 1 in a uint8_t, so LL=0xff (256-byte
op) wrapped to 0; alu_mvc's len/loop counter were also uint8_t. The
core-memory test (step 0x62, MVC 256, 0x000(0), 0x1500) therefore copied
ZERO bytes, so the fill never landed and the verify failed. Widen EXEC_SS
len and alu_mvc (param + loop index) to 16-bit.

With this the memory test fills + verifies and the deck runs steps 0x62/
0x63 + further memory phases, reaching report_and_end (0x19C6). Suite 243/245.

+++ alu_logic.c
@@ -28,9 +28,9 @@
 /* MVC – Move Characters                                               */
 /* ------------------------------------------------------------------ */
 
-void alu_mvc(struct ge *ge, uint16_t dst, uint16_t src, uint8_t len)
+void alu_mvc(struct ge *ge, uint16_t dst, uint16_t src, uint16_t len)
 {
-    uint8_t i;
+    uint16_t i;
     /*
      * The manual says "movement is from left to right through each field
      * a byte at a time" (§5.5.3.1).  A plain left-to-right byte loop is
